ConDec-375: Change tags in issue comments to JIRA like tags (#50)
* ConDec-375: Implement macros on client side. Failing tests disabled

* ConDec-375: Make Testcases great again. Implement new {tags} on client side.

* ConDec-375: Fix bugs from system tests

* ConDec-375: Move Mutex from Knowledge Rest to DecXtract Event handler

* ConDec-375: Fix codacy issues, Apply check if Decxtract and icon parsing are enabled. Implement icon parsing for event listener

* ConDec-375: Fixed indents and settings page issue for non settable toggles

* ConDec-375: write tags from automatic classfication back into comments

* ConDec-375: fix tree viewer in tab panel

* ConDec-375: Fix icons in TreeViewer

* ConDec-375: Add Project Check for DecXtract EL

* ConDEc-375: Fix Argument type icons in Treant

* ConDec-375 Fix codecy issues

+++ src/main/java/de/uhd/ifi/se/decision/management/jira/extraction/DecXtractEventListener.java
@@ -7,14 +7,18 @@ import org.springframework.stereotype.Component;
 
 import com.atlassian.event.api.EventListener;
 import com.atlassian.event.api.EventPublisher;
+import com.atlassian.jira.component.ComponentAccessor;
 import com.atlassian.jira.event.issue.IssueEvent;
 import com.atlassian.jira.event.type.EventType;
+import com.atlassian.jira.issue.comments.MutableComment;
 import com.atlassian.plugin.spring.scanner.annotation.imports.JiraImport;
 
 import de.uhd.ifi.se.decision.management.jira.extraction.connector.ViewConnector;
 import de.uhd.ifi.se.decision.management.jira.extraction.model.impl.CommentImpl;
+import de.uhd.ifi.se.decision.management.jira.extraction.model.util.CommentSplitter;
 import de.uhd.ifi.se.decision.management.jira.extraction.persistence.ActiveObjectsManager;
 import de.uhd.ifi.se.decision.management.jira.model.DecisionKnowledgeElementImpl;
+import de.uhd.ifi.se.decision.management.jira.persistence.ConfigPersistence;
 
 /**
  * Triggers the decXtract related function when some changes to comments are
@@ -29,6 +33,11 @@ public class DecXtractEventListener implements InitializingBean, DisposableBean 
 	private String projectKey;
 
 	private IssueEvent issueEvent;
+
+	/**
+	 * locks the editComment Event function if a rest service writes edites comments
+	 */
+	public static boolean editCommentLock;
 
 	@Autowired
 	public DecXtractEventListener(@JiraImport EventPublisher eventPublisher) {
@@ -59,12 +68,19 @@ public class DecXtractEventListener implements InitializingBean, DisposableBean 
 	public void onIssueEvent(IssueEvent issueEvent) {
 		this.issueEvent = issueEvent;
 		this.projectKey = issueEvent.getProject().getKey();
+		if(ConfigPersistence.isActivated(this.projectKey) && ConfigPersistence.isKnowledgeExtractedFromIssues(this.projectKey)) {
+			return;
+		}
 		Long eventTypeId = issueEvent.getEventTypeId();
+		if(eventTypeId.equals(EventType.ISSUE_COMMENTED_ID) || eventTypeId.equals(EventType.ISSUE_COMMENT_EDITED_ID)) {
+			parseIconsToTags();
+		}
 		if (eventTypeId.equals(EventType.ISSUE_COMMENTED_ID)) {
 			handleNewComment(new DecisionKnowledgeElementImpl(issueEvent.getIssue()));
 		}
 		if (eventTypeId.equals(EventType.ISSUE_COMMENT_DELETED_ID)) {
 			handleDeleteComment(new DecisionKnowledgeElementImpl(issueEvent.getIssue()));
+
 		}
 		if (eventTypeId.equals(EventType.ISSUE_COMMENT_EDITED_ID)) {
 			handleEditComment(new DecisionKnowledgeElementImpl(issueEvent.getIssue()));
@@ -72,17 +88,55 @@ public class DecXtractEventListener implements InitializingBean, DisposableBean 
 		if (eventTypeId.equals(EventType.ISSUE_DELETED_ID)) {
 			handleDeleteIssue(new DecisionKnowledgeElementImpl(issueEvent.getIssue()));
 		}
+		pareseSlashTagsOutOfComment();
+	}
+
+	private void parseIconsToTags() {
+		if(!ConfigPersistence.isIconParsing(issueEvent.getProject().getKey())) {
+			return ;
+		}
+		MutableComment comment = getCurrentEditedComment();
+		for (int i = 0; i < CommentSplitter.manualRationalIconList.length; i++) {
+			String icon = CommentSplitter.manualRationalIconList[i];			
+			while(comment.getBody().contains(icon)) {
+				comment.setBody(comment.getBody().replaceFirst(icon.replace("(", "\\(").replace(")", "\\)"), CommentSplitter.manualRationaleTagList[i]));
+				if(comment.getBody().split(System.getProperty("line.separator")).length == 1 && !comment.getBody().endsWith("\r\n")) {
+					comment.setBody(comment.getBody()+ CommentSplitter.manualRationaleTagList[i]);
+				}
+				comment.setBody(comment.getBody().replaceFirst("\r\n", CommentSplitter.manualRationaleTagList[i]));
+				ComponentAccessor.getCommentManager().update(comment, true);
+			}
+		}
+	}
+
+	/**
+	 * Parse \{ out of tags. These slashes are inserted by the visual mode editor
+	 */
+	private void pareseSlashTagsOutOfComment() {
+		if (issueEvent.getComment() != null && issueEvent.getComment().getBody().contains("\\{")) {
+			MutableComment comment = getCurrentEditedComment();
+			comment.setBody(issueEvent.getComment().getBody().replace("\\{", "{"));
+			ComponentAccessor.getCommentManager().update(comment, true);
+		}
 	}
 	
+	private MutableComment getCurrentEditedComment() {
+		return (MutableComment) ComponentAccessor.getCommentManager()
+				.getCommentById(issueEvent.getComment().getId());
+	}
+
 	private void handleDeleteIssue(DecisionKnowledgeElementImpl decisionKnowledgeElement) {
 		ActiveObjectsManager.cleanSentenceDatabaseForProject(this.projectKey);
 		ActiveObjectsManager.createLinksForNonLinkedElementsForIssue(decisionKnowledgeElement.getId() + "");
 	}
 
 	private void handleEditComment(DecisionKnowledgeElementImpl decisionKnowledgeElement) {
-		ActiveObjectsManager.checkIfCommentBodyHasChangedOutsideOfPlugin(new CommentImpl(issueEvent.getComment()));
-		new ViewConnector(this.issueEvent.getIssue(), false);
-		ActiveObjectsManager.createLinksForNonLinkedElementsForIssue(decisionKnowledgeElement.getId() + "");
+		if (!DecXtractEventListener.editCommentLock) {// If locked, a rest service is manipulating the comment and
+														// should not be handled by EL
+			ActiveObjectsManager.checkIfCommentBodyHasChangedOutsideOfPlugin(new CommentImpl(issueEvent.getComment()));
+			new ViewConnector(this.issueEvent.getIssue(), false);
+			ActiveObjectsManager.createLinksForNonLinkedElementsForIssue(decisionKnowledgeElement.getId() + "");
+		}
 	}
 
 	private void handleDeleteComment(DecisionKnowledgeElementImpl decisionKnowledgeElement) {
